+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Added
 - `Tuple6` through `Tuple8`
+- `Fn3#fn3` and `Fn4#fn4` static factory methods
 
 ## [2.0.0] - 2017-11-13
 ### Changed